Move components into relevant directories
