+++ web/src/components/cards/SmallCard.vue
