+++ web/src/components/cards/SuggestionCard.vue
