+++ web/src/components/carousels/StoryCarousel.vue
