+++ web/src/components/modals/CommentModal.vue
@@ -306,7 +306,7 @@ import { defineComponent, onMounted, computed } from 'vue';
 
 import type { PostCommentModal } from '@/common/models/post.model';
 import SVGLoader from '@/components/basics/SVGLoader.vue';
-import MediaCarousel from './MediaCarousel.vue';
+import MediaCarousel from '@/components/carousels/MediaCarousel.vue';
 
 
 export default defineComponent({
